==> 08_train_rnn_classifier.ipynb <==
import cv2
import glob
import ntpath
import random
import numpy

from keras.models import Model
from keras.optimizers import Adam
from keras.metrics import binary_accuracy, binary_crossentropy
from keras.callbacks import ModelCheckpoint, LearningRateScheduler
from keras.layers import Input, GRU, Dense, multiply, Flatten, TimeDistributed
from keras.layers.core import Permute

batch_size = 32

trainset, testset = load_data()

train_gen = data_generator(batch_size, trainset, need_augment=True)

test_gen = data_generator(batch_size, testset, need_augment=False)

learnrate_scheduler = LearningRateScheduler(step_decay)

model = get_net(load_weight_path=None)

checkpoint = ModelCheckpoint("G:/LungCancerPredict/model/" + "{epoch:02d}-{val_loss:.4f}.hd5", monitor='val_loss', period=1)

model.fit_generator(train_gen, int(len(trainset)/batch_size), 10, validation_data=test_gen, nb_val_samples=int(len(testset)/batch_size), callbacks=[checkpoint, learnrate_scheduler])

model.save("G:/LungCancerPredict/model/the_end.hd5")

def load_data(test_ratio=20, neg_per_pos=20):
    
    #load positive
    pos_samples = glob.glob("G:/LungCancerPredict/generated/luna16_positive_cube/*.png")
    random.shuffle(pos_samples)

    #load negative
    neg_samples_tn = glob.glob("G:/LungCancerPredict/generated/luna16_negative_cube/*_tn.png")
    neg_samples_fp = glob.glob("G:/LungCancerPredict/generated/luna16_negative_cube/*_fp.png")
    neg_samples_candidates = glob.glob("G:/LungCancerPredict/generated/luna16_negative_cube/*_candidates.png")
    neg_samples = neg_samples_fp + neg_samples_tn + neg_samples_candidates + neg_samples_candidates + neg_samples_candidates
    random.shuffle(neg_samples)

    # 正负样本混合
    pos_index = 0
    dataset = []
    for index, neg_sample in enumerate(neg_samples):
        
        dataset.append((neg_sample, 0, 0))
        
        if index % neg_per_pos == 0:
            
            pos_sample = pos_samples[pos_index]
            
            file_name = ntpath.basename(pos_sample)
            
            parts = file_name.split("_")
            class_label = int(parts[-2])
            size_label = int(parts[-3])
            
            dataset.append((pos_sample, class_label, size_label))
            pos_index += 1
            pos_index %= len(pos_samples)

    random.shuffle(dataset)
    counter = int(len(dataset) * ((100-test_ratio) / 100.0))
    dataset_train = dataset
    dataset_test = dataset[counter:]
                
    print("Train count: ", len(dataset_train), ", holdout count: ", len(dataset_test))
    
    return dataset_train, dataset_test

def data_generator(batch_size, record_list, need_augment):

    random_state = numpy.random.RandomState(999)
    
    if need_augment: random.shuffle(record_list)

    #按照batch_size动态生成数据
    batch_counter, crop_size, img_list, class_list, size_list = 0, 32, [], [], []
        
    for index, record in enumerate(record_list):
        
        class_label, size_label = record[1], record[2]
        
        if class_label == 0:
            
            cube_image = load_cube_img(record[0], 6, 8, 48)

            indent_x, indent_y, indent_z = 0, 0, 0
          
            wiggle = 48 - crop_size - 1

            indent_x, indent_y, indent_z = (random.randint(0, wiggle) for _ in range(3))

            cube_image = cube_image[indent_z:indent_z + crop_size, indent_y:indent_y + crop_size, indent_x:indent_x + crop_size]
        
            if need_augment:   
                if random.randint(0, 100) > 50: cube_image = numpy.fliplr(cube_image)
                if random.randint(0, 100) > 50: cube_image = numpy.flipud(cube_image)
                if random.randint(0, 100) > 50: cube_image = cube_image[:, :, ::-1]
                if random.randint(0, 100) > 50: cube_image = cube_image[:, ::-1, :]

        if class_label == 1:
            
            cube_image = load_cube_img(record[0], 8, 8, 64)
            
            if need_augment:
                wiggle_indent = crop_size / 4
                wiggle = 64 - crop_size - crop_size / 2 - 1
                indent_x = int(wiggle_indent + random.randint(0, wiggle))
                indent_y = int(wiggle_indent + random.randint(0, wiggle))
                indent_z = int(wiggle_indent + random.randint(0, wiggle))
            else:
                indent_x = int((64 - crop_size) / 2)
                indent_y = int((64 - crop_size) / 2)
                indent_z = int((64 - crop_size) / 2)

            cube_image = cube_image[indent_z:indent_z + crop_size, indent_y:indent_y + crop_size, indent_x:indent_x + crop_size]

            if need_augment:
                if random.randint(0, 100) > 50: cube_image = numpy.fliplr(cube_image)
                if random.randint(0, 100) > 50: cube_image = numpy.flipud(cube_image)
                if random.randint(0, 100) > 50: cube_image = cube_image[:, :, ::-1]
                if random.randint(0, 100) > 50: cube_image = cube_image[:, ::-1, :]
                    
        img3d = normalize_and_expand(cube_image)

        img_list.append(img3d), class_list.append(class_label), size_list.append(size_label)

        batch_counter += 1
        
        if batch_counter >= batch_size:
            
            x, y_class, y_size = numpy.vstack(img_list), numpy.vstack(class_list), numpy.vstack(size_list)

            yield x, {"out_class": y_class, "out_malignancy": y_size}
            
            img_list, class_list, size_list = [], [], []
            
            batch_counter = 0

def load_cube_img(src_path, rows, cols, size):
    
    img = cv2.imread(src_path, cv2.IMREAD_GRAYSCALE)
    res = numpy.zeros((rows * cols, size, size))

    img_height = size
    img_width = size

    for row in range(rows):
        
        for col in range(cols):
            
            src_y = row * img_height
            src_x = col * img_width
            res[row * cols + col] = img[src_y:src_y + img_height, src_x:src_x + img_width]

    return res

def normalize_and_expand(img):
    
    img = img.astype(numpy.float32)
    img -= 41 # 41 is MEAN_PIXEL_VALUE
    img /= 255.
    
    imgN=numpy.zeros(shape=(16,32,32))

    for i in range(16):
    
        imgN[i]=(img[2*i]+img[2*i+1])/2
        
    imgN = imgN.reshape(1,8,8,8,32)
    
    return imgN

def step_decay(epoch):
    
    res = 0.001
    
    if epoch > 5:
        
        res = 0.0001
        
    print("learnrate: ", res, " epoch: ", epoch)
    
    return res

def get_net(load_weight_path=None) -> Model:
        
    input_shapeT1=(8,32)  
    input1 = Input( shape=input_shapeT1 ) 
    gru1 = GRU(128,activation='tanh', recurrent_activation='sigmoid',return_sequences=True)(input1)
    gru1 = attention_block(gru1)
    gru1 = Flatten()(gru1)
    Encoder1 = Model(input1, gru1)

    input_shapeT2=(8,8,32)
    input2 = Input( shape=input_shapeT2 )
    embed2 = TimeDistributed(Encoder1)(input2)
    gru2 = GRU(256,activation='tanh', recurrent_activation='sigmoid',return_sequences=True)(embed2)
    gru2 = attention_block(gru2)
    gru2 = Flatten()(gru2)
    Encoder2 = Model(input2, gru2)
    
    input_shapeT3=(8,8,8,32) 
    input3 = Input( shape=input_shapeT3 )
    embed3 = TimeDistributed(Encoder2)(input3)
    gru3 = GRU(512,activation='tanh', recurrent_activation='sigmoid',return_sequences=True)(embed3)
    gru3 = attention_block(gru3)
    gru3 = Flatten()(gru3)
   
    gru3 = Dense(512, activation='relu')(gru3)
    out_class = Dense(1, activation='sigmoid', name='out_class')(gru3)
    model = Model(input=input3, output=out_class)
  
    if load_weight_path is not None: model.load_weights(load_weight_path, by_name=False)

    optimizer = Adam(lr=0.001, beta_1=0.9, beta_2=0.9, epsilon=1e-08, amsgrad=True)
    loss = {"out_class": "binary_crossentropy"}
    metrics = {"out_class": [binary_accuracy, binary_crossentropy]}
    
    model.compile(optimizer=optimizer, loss=loss, metrics=metrics)
    model.summary(line_length=140)
    
    return model

def attention_block(inputs):
    
    a = Permute((2, 1))(inputs)
    
    a = Dense(8, activation='softmax')(a)
    
    a_probs = Permute((2, 1), name='attention_vec')(a)
   
    output_attention_mul = multiply([inputs, a_probs], name='attention_mul')
    
    return output_attention_mul
