==> 18_make_feature_table_.ipynb <==
import pandas

make_feture_table()

def make_feture_table():

    df_labels = pandas.read_csv("G:/LungCancerPredict/resources/stage1_labels.csv")

    res_list = []
    
    for index, row in df_labels.iterrows():

        print(index)

        patient_id = row["id"]
        cancer_label = row["cancer"]
        
        df = [
            pandas.read_csv("G:/LungCancerPredict/detected/rnn_detect_ndsb3_nms_cube_attribute/" + patient_id + ".csv"),
            pandas.read_csv("G:/LungCancerPredict/detected/cnn_detect_ndsb3_nms_cube_attribute/" + patient_id + ".csv")
        ]
        df = pandas.concat(df, ignore_index=True)

        df = df[df["diamter"] > 0]
        df = df[df["nodule_chance"] > 0.5]
  
        if len(df) > 0:
            max_index = df["malignancy"].argmax()
            selected_nodule = df.loc[max_index] 
            res_list.append([cancer_label]+selected_nodule.values.tolist())
        else:
            res_list.append([cancer_label]+[patient_id , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 ])

    columns = ["cancer_label","patient_id","coord_x","coord_y","coord_z","nodule_chance","diamter","malignancy","spiculation","lobulation","sphericiy","margin","subtley","textrue","calcification"]

    df = pandas.DataFrame(res_list, columns=columns)

    df.to_csv("G:/LungCancerPredict/detected/feature_table.csv", index=False)
